Extracted hint functions in their own module and fixed issues with them
There is now a hint module.
Fixed pointers being passed to SDL.
Changed set_hint_with_priority to use a struct as priority.
Get_hint returns Option<String>
